Add example programs for XLOG language features and use cases
- Introduced a comprehensive suite of `.xlog` example programs organized by categories such as basics, arithmetic, graphs, security, supply chain, program analysis, database-style queries, and aggregates.
- Each example demonstrates specific language features, including transitive closure, stratified negation, constraints, comparisons, arithmetic operations, graph queries, RBAC, BOM explosion, and aggregation techniques.
- Included negative test cases to validate error handling for cycles, unstratifiable negation, type mismatches, and binding issues.
- Enhanced documentation with a detailed index and instructions on how to run the examples using the XLOG runner.

+++ crates/xlog-solve/benches/solver_bench.rs
@@ -417,7 +417,8 @@ fn bench_literal_operations(c: &mut Criterion) {
     });
 
     group.bench_function("to_packed_batch", |b| {
-        b.iter(|| literals.iter().map(|lit| lit.to_packed()).sum::<u32>());
+        // Use a wider accumulator so debug builds don't panic on overflow.
+        b.iter(|| literals.iter().map(|lit| lit.to_packed() as u64).sum::<u64>());
     });
 
     group.bench_function("from_packed_batch", |b| {
@@ -431,7 +432,8 @@ fn bench_literal_operations(c: &mut Criterion) {
     });
 
     group.bench_function("to_dimacs_batch", |b| {
-        b.iter(|| literals.iter().map(|lit| lit.to_dimacs()).sum::<i32>());
+        // Use a wider accumulator so debug builds don't panic on overflow.
+        b.iter(|| literals.iter().map(|lit| lit.to_dimacs() as i64).sum::<i64>());
     });
 
     group.finish();
